all code moved out of ref_resolve.rs
